SEMANA07/dia3/pos_backend/requirements.txt
django==3.2
djangorestframework
mysqlclient